调整缓存条目定义。添加异常enum定义。对话信息获取。房间信息获取。删除重复BusinessException。token校验和获取。

+++ peanut-chat-app-domain/src/main/java/com/sici/chat/model/chat/message/bo/MessageExtra.java
@@ -18,7 +18,7 @@ public class MessageExtra {
     /**
      * at的用户id列表
      */
-    private List<Integer> atUidList;
+    private List<Long> atUidList;
     /**
      * 消息撤回相关信息
      */
