+++ peanut-chat-app-domain/src/main/java/com/sici/chat/model/chat/message/vo/MessageVo.java
@@ -25,9 +25,9 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MessageVo {
-    private Integer id;
-    private Integer fromUid;
-    private Integer roomId;
+    private Long id;
+    private Long fromUid;
+    private Long roomId;
     private Integer type;
     private String content;
     private Date sendTime;
